Day 10.2: Pass tests but ...
I am dealing with a degree of complexity which I am not sure
if I can brute force it. Probably not given the previous exercises.

My best attempt now is to bump the number of max-clicks and hope
that all lines fit within that.

And hope that my computer can process them.